+++ nmnb-application/src/main/kotlin/nmnb/application/user/service/UserService.kt
@@ -1,0 +1,8 @@
+package nmnb.application.user.service
+
+import nmnb.application.user.service.dto.response.UserProfileResponse
+import nmnb.domain.user.User
+
+interface UserService {
+    fun getProfile(user: User): UserProfileResponse
+}
