+++ nmnb-application/src/test/kotlin/nmnb/application/ControllerTestSupport.kt
@@ -1,10 +1,11 @@
-package cv.nmnb
+package nmnb.application
 
-import cv.nmnb.domain.user.controller.UserController
-import cv.nmnb.domain.user.service.UserService
+import nmnb.application.user.controller.UserController
+import nmnb.application.user.service.UserService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.boot.test.mock.mockito.MockBean
+import org.springframework.test.context.ContextConfiguration
 import org.springframework.test.web.servlet.MockMvc
 
 @WebMvcTest(
@@ -12,6 +13,7 @@ import org.springframework.test.web.servlet.MockMvc
         UserController::class,
     ],
 )
+@ContextConfiguration(classes = [TestApplication::class])
 abstract class ControllerTestSupport {
     @Autowired
     protected lateinit var mockMvc: MockMvc
