+++ nmnb-application/src/test/kotlin/nmnb/application/IntegrationTestSupport.kt
@@ -1,8 +1,8 @@
-package cv.nmnb
+package nmnb.application
 
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
 
-@ActiveProfiles("test")
-@SpringBootTest
+@ActiveProfiles("local")
+@SpringBootTest(classes = [TestApplication::class])
 abstract class IntegrationTestSupport
