--- src/main/kotlin/cv/nmnb/domain/user/service/UserService.kt
@@ -1,8 +1,0 @@
-package cv.nmnb.domain.user.service
-
-import cv.nmnb.domain.user.domain.User
-import cv.nmnb.domain.user.service.dto.response.UserProfileResponse
-
-interface UserService {
-    fun getProfile(user: User): UserProfileResponse
-}
